not regard discriminator schema as a resolvedRef
Don't regard the discriminator schema as a resolvedRef, neither during
expanding deep levels nor construct a new schema. Instead, during
expanding the deep levels of a discriminator schema, regard the derived
schemas as the resolvedRef(s).

This is because in fact, the discriminator schema is not used at all, it
only works as a base and virtual class for serving the derived schemas.